Features: Pesquisa de aluno por inep_id; Alterar botão de próximo para salvar na edição de aluno; Adicionar o inep_id dos alunos nos relatórios; Gerenciar acessos (opção de ativar/desativar); (#146)
* Student search by inep_id

* Changed next button to save in student edit

Pending: In StudentController on line 265 there is an error in the condition (commented for testing purposes) that was preventing the student registration update

* Fixed error

In StudentController at line 265 and line 177 there is an error in the condition that was preventing the updating and creation of the student's record

* Class module reports with Inep ID information added

* Reports of students participating in Bolsa Família and students with pending documents with Inep ID added

* Added deactivate user function

* Console.log remove

* feat: Active and disable User

/app/views/activeDisableUser.php
/app/controllers/AdminController.php

* feature: Blocking mechanism for student duplicity; (CPF -> Certificate -> Name) :sparkles:

* feature: Error typing existing name in add student form :sparkles:

* feature: Error typing the existing cpf in the student addition form :sparkles:

* fix: added new name validation function for mother and father name fields :ambulance:

* feature: Error when typing the existing certificate number in the student inclusion form :sparkles:

* Bugfix: disable and enable pending buttons on icons :bug:

* Bugfix: missing new student name validation in another js file :bug:

* feature: display record similar to the student name entered in the student registration :sparkles:

* bug fix: re-enabled next button function and create/save button only appears on last registration page :bug:

* Update instance.php

* fix: Merge updates

* Update instance.php

* Update main.php

* feature: Able to change values ​​from previous years with the same login without having to access the year :sparkles:

* Update global.js

* Update StudentByClassroomReport.php

* Removed a modified file from pull request

* Removed a modified file from pull request

* refactor: added toggle and toggle icons :art:

* fix: Form errors now prevent student registration :bug:

* fix: temporary solution while not applying sass :lipstick:

+++ app/controllers/StudentController.php
@@ -38,7 +38,7 @@ class StudentController extends Controller
     {
         return array(
             array('allow', // allow authenticated user to perform 'create' and 'update' actions
-                'actions' => array('index', 'view', 'create', 'update', 'getcities', 'getnotaryoffice', 'getnations', 'delete'),
+                'actions' => array('index', 'view', 'comparestudentname', 'comparestudentcpf', 'comparestudentcertificate', 'create', 'update', 'getcities', 'getnotaryoffice', 'getnations', 'delete'),
                 'users' => array('@'),
             ),
             array('deny', // deny all users
@@ -121,6 +121,31 @@ class StudentController extends Controller
         }
     }
 
+    public function actionCompareStudentName() { 
+        $data = StudentIdentification::model()->findAll();
+        $result = [];
+        foreach ($data as $student) {
+            $result[$student->name] = $student->id;
+        }
+        echo json_encode($result);
+    }
+
+    public function actionCompareStudentCertificate($student_certificate) {
+        $data = StudentDocumentsAndAddress::model()->find('civil_certification_term_number=:civil_certification_term_number', array(':civil_certification_term_number' => $student_certificate));
+        $result = [];
+        $result[$data->student_fk] = $data->id;
+
+        echo json_encode($result);
+    }
+
+    public function actionCompareStudentCpf($student_responsable_cpf) {
+        $data = StudentIdentification::model()->find('responsable_cpf=:responsable_cpf', array(':responsable_cpf' => $student_responsable_cpf));
+        $result = [];
+        $result[$data->name] = $data->id;
+
+        echo json_encode($result);
+    }
+
 
     /**
      * Creates a new model.
@@ -135,7 +160,7 @@ class StudentController extends Controller
         $modelEnrollment = new StudentEnrollment;
 
         $vaccines = Vaccine::model()->findAll(array('order' => 'name'));
-        $studentVaccinesSaves = StudentVaccine::model()->findAll(['select' => 'vaccine_id', 'condition' => 'student_id=:student_id', 'params' => [':student_id' => $id]]);
+        $studentVaccinesSaves = StudentVaccine::model()->findAll(['select' => 'vaccine_id', 'condition' => 'student_id=:student_id', 'params' => [':student_id' => $modelStudentIdentification->id]]);
         if ($studentVaccinesSaves) {
             $studentVaccinesSaves = array_map(function ($item) {
                 return $item->vaccine_id;
@@ -147,6 +172,27 @@ class StudentController extends Controller
         if (isset($_POST[$this->STUDENT_IDENTIFICATION]) && isset($_POST[$this->STUDENT_DOCUMENTS_AND_ADDRESS])) {
             $modelStudentIdentification->attributes = $_POST[$this->STUDENT_IDENTIFICATION];
             $modelStudentDocumentsAndAddress->attributes = $_POST[$this->STUDENT_DOCUMENTS_AND_ADDRESS];
+
+            // Validação CPF->Certidão->Nome
+            if($modelStudentIdentification->responsable_cpf != null) {
+                $student_test_cpf = StudentIdentification::model()->find('responsable_cpf=:responsable_cpf', array(':responsable_cpf'=>$modelStudentIdentification->responsable_cpf));
+                if(isset($student_test_cpf)) {
+                    Yii::app()->user->setFlash('error', Yii::t('default', "O CPF do responsável informado já está cadastrado"));
+                    $this->redirect(array('index'));
+                }
+            }
+            if($modelStudentDocumentsAndAddress->civil_certification_term_number != null) {
+                $student_test_certificate = StudentIdentification::model()->find('civil_certification_term_number=:civil_certification_term_number', array(':civil_certification_term_number'=>$modelStudentDocumentsAndAddress->civil_certification_term_number));
+                if(isset($student_test_certificate)) {
+                    Yii::app()->user->setFlash('error', Yii::t('default', "O Nº do Termo da Certidão informado já está cadastrado"));
+                    $this->redirect(array('index'));
+                }
+            }
+            $student_test_name = StudentIdentification::model()->find('name=:name', array(':name'=>$modelStudentIdentification->name));
+            if(isset($student_test_name)) {
+                Yii::app()->user->setFlash('error', Yii::t('default', "O nome do aluno informado já está cadastrado"));
+                $this->redirect(array('index'));
+            }
 
             //Atributos comuns entre as tabelas
             $modelStudentDocumentsAndAddress->school_inep_id_fk = $modelStudentIdentification->school_inep_id_fk;
@@ -174,14 +220,16 @@ class StudentController extends Controller
                                 //$modelEnrollment = $this->loadModel($id, $this->STUDENT_ENROLLMENT);
                             }
 
-                            if (count($_POST['Vaccine']['vaccine_id']) > 0) {
-                                StudentVaccine::model()->deleteAll("student_id = $modelStudentIdentification->id");
-
-                                foreach ($_POST['Vaccine']['vaccine_id'] as $vaccine_id) {
-                                    $studentVaccine = new StudentVaccine();
-                                    $studentVaccine->student_id = $modelStudentIdentification->id;
-                                    $studentVaccine->vaccine_id = $vaccine_id;
-                                    $studentVaccine->save();
+                            if(isset($_POST['Vaccine']['vaccine_id'])) {
+                                if (count($_POST['Vaccine']['vaccine_id']) > 0) {
+                                    StudentVaccine::model()->deleteAll("student_id = $modelStudentIdentification->id");
+    
+                                    foreach ($_POST['Vaccine']['vaccine_id'] as $vaccine_id) {
+                                        $studentVaccine = new StudentVaccine();
+                                        $studentVaccine->student_id = $modelStudentIdentification->id;
+                                        $studentVaccine->vaccine_id = $vaccine_id;
+                                        $studentVaccine->save();
+                                    }
                                 }
                             }
 
@@ -262,16 +310,18 @@ class StudentController extends Controller
                             //$modelEnrollment = $this->loadModel($id, $this->STUDENT_ENROLLMENT);
                         }
 
-                        if (count($_POST['Vaccine']['vaccine_id']) > 0) {
-                            if ($studentVaccinesSaves) {
-                                StudentVaccine::model()->deleteAll("student_id = $modelStudentIdentification->id");
-                            }
+                        if(isset($_POST['Vaccine']['vaccine_id'])) {
+                            if (count($_POST['Vaccine']['vaccine_id']) > 0) {
+                                if ($studentVaccinesSaves) {
+                                    StudentVaccine::model()->deleteAll("student_id = $modelStudentIdentification->id");
+                                }
 
-                            foreach ($_POST['Vaccine']['vaccine_id'] as $vaccine_id) {
-                                $studentVaccine = new StudentVaccine();
-                                $studentVaccine->student_id = $modelStudentIdentification->id;
-                                $studentVaccine->vaccine_id = $vaccine_id;
-                                $studentVaccine->save();
+                                foreach ($_POST['Vaccine']['vaccine_id'] as $vaccine_id) {
+                                    $studentVaccine = new StudentVaccine();
+                                    $studentVaccine->student_id = $modelStudentIdentification->id;
+                                    $studentVaccine->vaccine_id = $vaccine_id;
+                                    $studentVaccine->save();
+                                }
                             }
                         }
